chore: update version

+++ links-soma-citygml-conv/nusamai/src/sink/gpkg/mod.rs
@@ -5,7 +5,6 @@ mod bbox;
 mod table;
 use attributes::prepare_object_attributes;
 use bbox::{get_indexed_multipolygon_bbox, Bbox};
-use chrono::Utc;
 use indexmap::IndexMap;
 use nusamai_citygml::{
     object::{ObjectStereotype, Value},
@@ -16,7 +15,6 @@ use nusamai_gpkg::{geometry::write_indexed_multipolygon, GpkgHandler};
 use rayon::prelude::*;
 use serde_json::Value as JsonValue;
 use std::collections::HashMap;
-use std::env;
 use std::{collections::HashSet, path::PathBuf, time::Duration};
 use table::schema_to_table_infos;
 use std::thread;
@@ -30,6 +28,10 @@ use crate::{
     transformer,
     transformer::TransformerRegistry,
 };
+use std::fs::{self, File};
+use std::io::{Write, BufReader, Read};
+use zip::{write::FileOptions, CompressionMethod, ZipWriter};
+use walkdir::WalkDir;
 
 use super::option::output_parameter;
 #[derive(Clone)]
@@ -200,6 +202,16 @@ impl GpkgSink {
         let mut merged_data: HashMap<String, FeatureData> = HashMap::new();
         let mut records = Vec::new();
         let mut handlers: HashMap<String, GpkgHandler> = HashMap::new();
+        let mut output_path_init = PathBuf::new();
+
+        let mut path_delete = PathBuf::from(&self.output_path);
+        path_delete = path_delete.join("PLATEAU.zip");
+
+        if path_delete.exists() {
+            if let Err(e) = fs::remove_file(&path_delete) {
+                eprintln!("Error deleting directory: {}", e);
+            }
+        }
 
         while let Some((table_name, record)) = receiver.recv().await {
             feedback.ensure_not_canceled()?;
@@ -207,43 +219,54 @@ impl GpkgSink {
             if !created_tables.contains(&table_name) {
                 let tf = table_infos.get(&table_name).unwrap();
                 if tf.has_geometry {
-                    let filename = table_name.replace(':', "_");
-                    let mut path = PathBuf::from(&self.output_path);
-                    if let Some(stem) = path.file_stem() {
-                        let new_filename = format!("{}.gpkg", filename);
-                        path = path.join(new_filename);
-                    }
+                    let filename = Self::convert_filename(&table_name);
+                    if filename != "PLATEAUurf" {
+                        let mut path = PathBuf::from(&self.output_path);
 
-                    let conn_str = format!("file:{}", path.to_string_lossy());
+                        if !path.exists() {
+                            if let Err(e) = fs::create_dir(&path) {
+                                eprintln!("Failed to create directory: {}", e);
+                            }
+                        }
+                        if output_path_init.as_os_str().is_empty() {
+                            output_path_init = path.clone();
+                        }
+                        if let Some(_stem) = path.file_stem() {
+                            let new_filename = format!("{}.gpkg", filename);
+                            path = path.join(new_filename);
+                        }
 
-                    if path.exists() {
-                        std::fs::remove_file(&path)?;
-                    };
+                        let conn_str = format!("file:{}", path.to_string_lossy());
 
-                    if !handlers.contains_key(&table_name) {
-                        let new_handler = GpkgHandler::from_str(&conn_str)
+                        if path.exists() {
+                            std::fs::remove_file(&path)?;
+                        };
+
+                        if !handlers.contains_key(&table_name) {
+                            let new_handler = GpkgHandler::from_str(&conn_str)
+                                .await
+                                .expect("Failed to connect GpkgHandler");
+
+                            handlers.insert(table_name.clone(), new_handler);
+                        }
+
+                        let handler = handlers.get_mut(&table_name).unwrap();
+
+                        let mut tx = handler
+                            .begin()
                             .await
-                            .expect("Failed to connect GpkgHandler");
+                            .map_err(|e| PipelineError::Other(e.to_string()))?;
 
-                        handlers.insert(table_name.clone(), new_handler);
+                        tx.add_table(tf, srs_id)
+                            .await
+                            .map_err(|e| PipelineError::Other(e.to_string()))?;
+
+                        tx.commit()
+                            .await
+                            .map_err(|e| PipelineError::Other(e.to_string()))?;
+
+                        created_tables.insert(table_name.clone());
                     }
-
-                    let handler = handlers.get_mut(&table_name).unwrap();
-
-                    let mut tx = handler
-                        .begin()
-                        .await
-                        .map_err(|e| PipelineError::Other(e.to_string()))?;
-
-                    tx.add_table(tf, srs_id)
-                        .await
-                        .map_err(|e| PipelineError::Other(e.to_string()))?;
-
-                    tx.commit()
-                        .await
-                        .map_err(|e| PipelineError::Other(e.to_string()))?;
-
-                    created_tables.insert(table_name.clone());
                 }
             }
 
@@ -298,7 +321,7 @@ impl GpkgSink {
                                 None
                             };
                     
-                            if let Some(cat) = category {
+                            if let Some(_cat) = category {
                                 if !matches!(
                                     table_name.as_str(),
                                     "uro:BuildingIDAttribute"
@@ -406,19 +429,14 @@ impl GpkgSink {
             tx.commit()
                 .await
                 .map_err(|e| PipelineError::Other(e.to_string()))?;
-
-            thread::sleep(Duration::from_millis(1000));
         }
 
-        let output_dir = env::temp_dir();
-        let timestamp = Self::get_timestamp();
-        let target_dir = output_dir.join(timestamp);
-        if target_dir.exists() {
-            if let Err(e) = std::fs::remove_dir_all(&target_dir) {
-                println!("Error delete directory: {:?}", e);
-            }
+        for handler in handlers.values() {
+            handler.close().await;
         }
-        thread::sleep(Duration::from_millis(1000));
+
+        let _ = Self::zip_folder(output_path_init);
+
         feedback.info("Complete processing".into());
 
         match producers.await.unwrap() {
@@ -427,9 +445,44 @@ impl GpkgSink {
         }
     }
 
-    fn get_timestamp() -> String {
-        let now = Utc::now();
-        format!("linksoma{}", now.format("%Y%m%d"))
+    fn zip_folder(src_folder: PathBuf) -> std::io::Result<()> {
+        let zip_file_path = src_folder.join("PLATEAU.zip");
+        let zip_file = File::create(&zip_file_path)?;
+        let mut zip = ZipWriter::new(zip_file);
+    
+        let options = FileOptions::default().compression_method(CompressionMethod::Deflated);
+        let buffer_size: usize = 5 * 1024 * 1024; // 5MB buffer
+        let sleep_time = Duration::from_millis(500);
+    
+        for entry in WalkDir::new(&src_folder).max_depth(1) {
+            let entry = entry?;
+            let entry_path = entry.path();
+    
+            if entry_path.is_file() && entry_path.extension().map_or(false, |ext| ext == "gpkg") {
+                let relative_path = entry_path.strip_prefix(&src_folder).unwrap();
+                zip.start_file(relative_path.to_string_lossy(), options)?;
+    
+                let mut file = BufReader::new(File::open(entry_path)?);
+                let mut buffer = vec![0; buffer_size];
+    
+                loop {
+                    let bytes_read = file.read(&mut buffer)?;
+                    if bytes_read == 0 {
+                        break;
+                    }
+                    zip.write_all(&buffer[..bytes_read])?;
+    
+                    if bytes_read >= buffer_size {
+                        thread::sleep(sleep_time);
+                    }
+                }
+    
+                fs::remove_file(entry_path)?;
+            }
+        }
+    
+        zip.finish()?; 
+        Ok(())
     }
 
     fn rename_all_attributes(
@@ -461,7 +514,7 @@ impl GpkgSink {
         let keys_to_rename: Vec<String> = updated_attributes.keys().cloned().collect();
     
         for key in keys_to_rename {
-            if let Some(value) = updated_attributes.remove(&key) {
+            if let Some(value) = updated_attributes.shift_remove(&key) {
                 let new_key = if new_prefix.is_empty() {
                     format!("{}|{}", new_table_name, key)
                 } else {
@@ -469,6 +522,112 @@ impl GpkgSink {
                 };
                 updated_attributes.insert(new_key, value);
             }
+        }
+    }
+
+    fn convert_filename(table_name: &str) -> String {
+        match table_name {
+            "bldg:Building" => "PLATEAU建築物".to_string(),
+            "luse:LandUse" => "PLATEAU土地利用".to_string(),
+            "urf:UrbanPlanningArea" => "PLATEAU都市計画区域".to_string(),
+            "urf:QuasiUrbanPlanningArea" => "PLATEAU準都市計画区域".to_string(),
+            "urf:AreaClassification" => "PLATEAU区域区分".to_string(),
+            "urf:DistrictsAndZones" => "PLATEAU地域地区".to_string(),
+            "urf:UseDistrict" => "PLATEAU用途地域".to_string(),
+            "urf:SpecialUseDistrict" => "PLATEAU特別用途地区".to_string(),
+            "urf:SpecialUseRestrictionDistrict" => "PLATEAU特定用途制限地域".to_string(),
+            "urf:ExceptionalFloorAreaRateDistrict" => "PLATEAU特例容積率適用地区".to_string(),
+            "urf:HighRiseResidentialAttractionDistrict" => "PLATEAU高層住居誘導地区".to_string(),
+            "urf:HeightControlDistrict" => "PLATEAU高度地区".to_string(),
+            "urf:HighLevelUseDistrict" => "PLATEAU高度利用地区".to_string(),
+            "urf:SpecifiedBlock" => "PLATEAU特定街区".to_string(),
+            "urf:SpecialUrbanRenaissanceDistrict" => "PLATEAU都市再生特別地区".to_string(),
+            "urf:HousingControlArea" => "PLATEAU居住調整地域".to_string(),
+            "urf:ResidentialEnvironmentImprovementDistrict" => "PLATEAU居住環境向上用途誘導地区".to_string(),
+            "urf:SpecialUseAttractionDistrict" => "PLATEAU特定用途誘導地区".to_string(),
+            "urf:FirePreventionDistrict" => "PLATEAU防火地域又は準防火地域".to_string(),
+            "urf:SpecifiedDisasterPreventionBlockImprovementZone" => "PLATEAU特定防災街区整備地区".to_string(),
+            "urf:LandscapeZone" => "PLATEAU景観地区".to_string(),
+            "urf:ScenicDistrict" => "PLATEAU風致地区".to_string(),
+            "urf:ParkingPlaceDevelopmentZone" => "PLATEAU駐車場整備地区".to_string(),
+            "urf:PortZone" => "PLATEAU臨港地区".to_string(),
+            "urf:SpecialZoneForPreservationOfHistoricalLandscape" => "PLATEAU歴史的風土特別保存地区".to_string(),
+            "urf:ZoneForPreservationOfHistoricalLandscape" => "PLATEAU風土保存地区".to_string(),
+            "urf:GreenSpaceConservationDistrict" => "PLATEAU緑地保全地域".to_string(),
+            "urf:SpecialGreenSpaceConservationDistrict" => "PLATEAU特別緑地保全地域".to_string(),
+            "urf:TreePlantingDistrict" => "PLATEAU緑化地域".to_string(),
+            "urf:DistributionBusinessZone" => "PLATEAU流通業務地区".to_string(),
+            "urf:ProductiveGreenZone" => "PLATEAU生産緑地地区".to_string(),
+            "urf:ConservationZoneForClustersOfTraditionalStructures" => "PLATEAU伝統的建造物群保存地区".to_string(),
+            "urf:AircraftNoiseControlZoneurf:AircraftNoiseControlZone" => "PLATEAU防止特別地区".to_string(),
+            "urf:ProjectPromotionArea" => "PLATEAU促進区域".to_string(),
+            "urf:UrbanRedevelopmentPromotionArea" => "PLATEAU市街地再開発促進区域".to_string(),
+            "urf:LandReadjustmentPromotionArea" => "PLATEAU土地区画整理促進区域".to_string(),
+            "urf:ResidentialBlockConstructionPromotionArea" => "PLATEAU住宅街区整備促進区域".to_string(),
+            "urf:LandReadjustmentPromotionAreasForCoreBusinessUrbanDevelopment" => "PLATEAU拠点業務市街地整備土地区画整理促進区域".to_string(),
+            "urf:UnusedLandUsePromotionArea" => "PLATEAU遊休土地転換利用促進地区".to_string(),
+            "urf:UrbanDisasterRecoveryPromotionArea" => "PLATEAU被災市街地復興推進地域".to_string(),
+            "urf:UrbanFacility" => "PLATEAU都市施設".to_string(),
+            "urf:TrafficFacility" => "PLATEAU交通施設".to_string(),
+            "urf:OpenSpaceForPublicUse" => "PLATEAU公共空地".to_string(),
+            "urf:SupplyFacility," => "PLATEAU供給施設及び処理施設".to_string(),
+            "urf:TreatmentFacility" => "PLATEAU供給施設及び処理施設".to_string(),
+            "urf:Waterway" => "PLATEAU水路".to_string(),
+            "urf:EducationalAndCulturalFacility" => "PLATEAU教育文化施設".to_string(),
+            "urf:MedicalFacility" => "PLATEAU医療施設及び社会福祉施設".to_string(),
+            "urf:SocialWelfareFacility" => "PLATEAU医療施設及び社会福祉施設".to_string(),
+            "urf:MarketsSlaughterhousesCrematoria" => "PLATEAU市場、と畜場、火葬場".to_string(),
+            "urf:CollectiveHousingFacilities" => "PLATEAU一団地の住宅施設".to_string(),
+            "urf:CollectiveGovernmentAndPublicOfficeFacilities" => "PLATEAU一団地の官公庁施設".to_string(),
+            "urf:DistributionBusinessPark" => "PLATEAU流通業務団地".to_string(),
+            "urf:CollectiveFacilitiesForTsunamiDisasterPrevention" => "PLATEAU一団地の津波防災拠点市街地形成施設".to_string(),
+            "urf:CollectiveFacilitiesForReconstructionAndRevitalization" => "PLATEAU一団地の復興再生拠点市街地形成施設".to_string(),
+            "urf:CollectiveFacilitiesForReconstruction" => "PLATEAU一団地の復興拠点市街地形成施設".to_string(),
+            "urf:CollectiveUrbanDisasterPreventionFacilities" => "PLATEAU一団地の都市安全確保拠点施設".to_string(),
+            "urf:UrbanFacilityStipulatedByCabinetOrder" => "PLATEAU政令で定める都市施設".to_string(),
+            "urf:TelecommunicationFacility" => "PLATEAU電気通信施設".to_string(),
+            "urf:WindProtectionFacility" => "PLATEAU防風施設".to_string(),
+            "urf:FireProtectionFacility" => "PLATEAU防火施設".to_string(),
+            "urf:TideFacility" => "PLATEAU防潮施設".to_string(),
+            "urf:FloodPreventionFacility" => "PLATEAU防水施設".to_string(),
+            "urf:SnowProtectionFacility" => "PLATEAU防雪施設".to_string(),
+            "urf:SandControlFacility" => "PLATEAU防砂施設".to_string(),
+            "urf:UrbanDevelopmentProject" => "PLATEAU市街地開発事業".to_string(),
+            "urf:LandReadjustmentProject" => "PLATEAU土地区画整理事業".to_string(),
+            "urf:NewHousingAndUrbanDevelopmentProject" => "PLATEAU新住宅市街地開発事業".to_string(),
+            "urf:IndustrialParkDevelopmentProject" => "PLATEAU工業団地造成事業".to_string(),
+            "urf:UrbanRedevelopmentProject" => "PLATEAU市街地再開発事業".to_string(),
+            "urf:NewUrbanInfrastructureProject" => "PLATEAU新都市基盤整備事業".to_string(),
+            "urf:ResidentialBlockConstructionProject" => "PLATEAU住宅街区整備事業".to_string(),
+            "urf:DisasterPreventionBlockImprovementProject" => "PLATEAU防災街区整備事業".to_string(),
+            "urf:UrbanRenewalProject" => "PLATEAU市街地改造事業".to_string(),
+            "urf:ScheduledAreaForUrbanDevelopmentProject" => "PLATEAU市街地開発事業等の予定区域".to_string(),
+            "urf:ScheduledAreaForNewHousingAndUrbanDevelopmentProjects" => "PLATEAU新住宅市街地開発事業の予定区域".to_string(),
+            "urf:ScheduledAreaForIndustrialParkDevelopmentProjects" => "PLATEAU工業団地造成事業の予定区域".to_string(),
+            "urf:ScheduledAreaForNewUrbanInfrastructureProjects" => "PLATEAU新都市基盤整備事業の予定区域".to_string(),
+            "urf:ScheduledAreaForCollectiveHousingFacilities" => "PLATEAU一団地の住宅施設の予定区域".to_string(),
+            "urf:ScheduledAreaForCollectiveGovernmentAndPublicOfficeFacilities" => "PLATEAU一団地の官公庁施設の予定区域".to_string(),
+            "urf:ScheduledAreaForDistributionBusinessPark" => "PLATEAU流通業務団地の予定区域".to_string(),
+            "urf:DistrictPlan" => "PLATEAU地区計画".to_string(),
+            "urf:DistrictDevelopmentPlan" => "PLATEAU地区整備計画".to_string(),
+            "urf:DistrictFacilityurf:DistrictFacility" => "PLATEAU地区施設".to_string(),
+            "urf:RoadsideDistrictPlan" => "PLATEAU沿道地区計画".to_string(),
+            "urf:RoadsideDistrictImprovementPlan" => "PLATEAU沿道地区整備計画".to_string(),
+            "urf:RoadsideDistrictFacility" => "PLATEAU沿道地区施設".to_string(),
+            "urf:RuralDistrictPlan" => "PLATEAU集落地区計画".to_string(),
+            "urf:RuralDistrictImprovementPlan" => "PLATEAU集落地整備計画".to_string(),
+            "urf:RuralDistrictFacility" => "PLATEAU集落施設".to_string(),
+            "urf:HistoricSceneryMaintenanceAndImprovementDistrictPlan" => "PLATEAU歴史的風致維持向上地区計画".to_string(),
+            "urf:DistrictImprovementPlanForHistoricSceneryMaintenanceAndImprovementDistrict" => "PLATEAU歴史的風致維持向上地区整備計画".to_string(),
+            "urf:DisasterPreventionBlockImprovementZonePlan" => "PLATEAU防災街区整備地区計画".to_string(),
+            "urf:SpecifiedBuildingZoneImprovementPlan" => "PLATEAU特定建築物地区整備計画".to_string(),
+            "urf:DistrictImprovementPlanForDisasterPreventionBlockImprovementZonePlan" => "PLATEAU防災街区整備地区整備計画".to_string(),
+            "urf:ZonalDisasterPreventionFacility" => "PLATEAU地区防災施設".to_string(),
+            "urf:ThreeDimensionalExtent" => "PLATEAU立体的な範囲".to_string(),
+            "urf:Boundary" => "PLATEAU境界".to_string(),
+            "urf:UrbanFunctionAttractionArea" => "PLATEAU都市機能誘導区域".to_string(),
+            "urf:ResidenceAttractionArea" => "PLATEAU居住誘導区域".to_string(),
+            _ => format!("PLATEAU{}", table_name),
         }
     }
